+++ src/components/SymbolModal.css
@@ -15,10 +15,6 @@
   top: 30px;
 }
 
-.theme-dark .symbol-modal {
-  box-shadow: 2px 4px 6px var(--popup-shadow-color);
-}
-
 .symbol-modal .input-wrapper,
 .modal .input-wrapper {
   display: flex;
